requirements.txt
#wheel
wheel==0.37.1
#devops
black==22.3.0
click==8.1.3 
pytest==7.1.3
pytest-cov==4.0.0
pylint==2.15.3
boto3==1.24.87
ipykernel
#web
fastapi==0.85.0
uvicorn==0.18.3
#datascience
jupyter==1.0.0
pandas==1.5.0
numpy==1.23.3
scikit-learn==1.1.2
matplotlib==3.6.0
seaborn==0.12.0
dask==2022.9.2
#geo
geopy==2.2.0
#MLOps
gradio==3.4.1
streamlit==1.13.0
protobuf==3.16.0
onnx==1.12.0
databricks==0.2
databricks-cli==0.17.3
databricks-sql-connector==2.1.0
##MLOps-HuggingFace
transformers==4.22.2
datasets==2.5.2
tokenizers==0.12.1
nltk==3.7
rouge-score==0.1.2
py7zr==0.20.0
evaluate[template]==0.3.0
#deep learning
#torch==1.12.1
torchvision==0.13.1
#sql 
google-cloud-bigquery
db-dtypes
